Moved VersionHandler and implementations of VersionHandler into its own sub-package
