+++ Reflection/src/main/java/me/doclic/noencryption/compatibility/versionhandler/VersionHandler11900.java
@@ -1,5 +1,6 @@
-package me.doclic.noencryption.compatibility;
+package me.doclic.noencryption.compatibility.versionhandler;
 
+import me.doclic.noencryption.compatibility.NMSInterface;
 import me.doclic.noencryption.utils.InternalMetrics;
 import me.doclic.noencryption.utils.Metrics;
